# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Starwars",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return a query",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a registered query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Angel",
                                    "email": "angelmserranog@gmail.com",
                                    "password": 12345678
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns a updated query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete to data",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a delete query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Query all vehicles",
                "operationId": "indexVehicles",
                "responses": {
                    "200": {
                        "description": "returns the registered vehicles"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "register vehicle, only place 6 fields but 16 can be registered, this is just an example",
                "operationId": "storeVehicle",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "manufacturer": {
                                        "type": "string"
                                    },
                                    "cost_in_credits": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "string"
                                    },
                                    "max_atmosphering_speed": {
                                        "type": "string"
                                    },
                                    "starship_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Sand Angel",
                                    "model": "Digger Crawler",
                                    "manufacturer": "Corellia Mining Corporation",
                                    "cost_in_credits": "75000",
                                    "length": "7",
                                    "max_atmosphering_speed": 1500,
                                    "starship_id": 5
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns the registered object"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Return a query",
                "operationId": "showVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a query query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "update vehicle, only place 6 fields but 16 can be update, this is just an example",
                "operationId": "updateVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "manufacturer": {
                                        "type": "string"
                                    },
                                    "cost_in_credits": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "string"
                                    },
                                    "max_atmosphering_speed": {
                                        "type": "string"
                                    },
                                    "starship_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Sand Angel",
                                    "model": "Digger Crawler",
                                    "manufacturer": "Corellia Mining Corporation",
                                    "cost_in_credits": "75000",
                                    "length": "7",
                                    "max_atmosphering_speed": 1500,
                                    "starship_id": 5
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns the update object"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete to data",
                "operationId": "destroyVehicles",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a delete query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}